+++ cppad/local/pow_op.hpp
@@ -151,8 +151,8 @@ inline void forward_powvv_op_0(
 	CPPAD_ASSERT_UNKNOWN( NumRes(PowvvOp) == 3 );
 
 	// Taylor coefficients corresponding to arguments and result
-	Base* x   = taylor + arg[0] * cap_order;
-	Base* y   = taylor + arg[1] * cap_order;
+	Base* x   = taylor + size_t(arg[0]) * cap_order;
+	Base* y   = taylor + size_t(arg[1]) * cap_order;
 	Base* z_0 = taylor + i_z    * cap_order;
 	Base* z_1 = z_0    +          cap_order;
 	Base* z_2 = z_1    +          cap_order;
@@ -385,7 +385,7 @@ inline void forward_powpv_op_0(
 	Base x = parameter[ arg[0] ];
 
 	// Taylor coefficients corresponding to arguments and result
-	Base* y   = taylor + arg[1] * cap_order;
+	Base* y   = taylor + size_t(arg[1]) * cap_order;
 	Base* z_0 = taylor + i_z    * cap_order;
 	Base* z_1 = z_0    +          cap_order;
 	Base* z_2 = z_1    +          cap_order;
@@ -508,7 +508,7 @@ inline void forward_powvp_op(
 	// zero order case exactly same as Base type operation
 	if( p == 0 )
 	{	Base* z_2 = taylor + (i_z+2) * cap_order;
-		Base* x   = taylor + arg[0] * cap_order;
+		Base* x   = taylor + size_t(arg[0]) * cap_order;
 		Base  y   = parameter[ arg[1] ];
 		z_2[0]  = pow(x[0], y);
 		p++;
@@ -596,7 +596,7 @@ inline void forward_powvp_op_0(
 	Base y = parameter[ arg[1] ];
 
 	// Taylor coefficients corresponding to arguments and result
-	Base* x   = taylor + arg[0] * cap_order;
+	Base* x   = taylor + size_t(arg[0]) * cap_order;
 	Base* z_0 = taylor + i_z    * cap_order;
 	Base* z_1 = z_0    +          cap_order;
 	Base* z_2 = z_1    +          cap_order;
