+++ cppad/local/sweep/forward1.hpp
@@ -337,7 +337,7 @@ void forward1(
 		switch( op )
 		{
 			case AbsOp:
-			forward_abs_op(p, q, i_var, arg[0], J, taylor);
+			forward_abs_op(p, q, i_var, size_t(arg[0]), J, taylor);
 			break;
 			// -------------------------------------------------
 
@@ -355,7 +355,7 @@ void forward1(
 			case AcosOp:
 			// sqrt(1 - x * x), acos(x)
 			CPPAD_ASSERT_UNKNOWN( i_var < numvar  );
-			forward_acos_op(p, q, i_var, arg[0], J, taylor);
+			forward_acos_op(p, q, i_var, size_t(arg[0]), J, taylor);
 			break;
 			// -------------------------------------------------
 
@@ -363,7 +363,7 @@ void forward1(
 			case AcoshOp:
 			// sqrt(x * x - 1), acosh(x)
 			CPPAD_ASSERT_UNKNOWN( i_var < numvar  );
-			forward_acosh_op(p, q, i_var, arg[0], J, taylor);
+			forward_acosh_op(p, q, i_var, size_t(arg[0]), J, taylor);
 			break;
 # endif
 			// -------------------------------------------------
@@ -371,7 +371,7 @@ void forward1(
 			case AsinOp:
 			// sqrt(1 - x * x), asin(x)
 			CPPAD_ASSERT_UNKNOWN( i_var < numvar  );
-			forward_asin_op(p, q, i_var, arg[0], J, taylor);
+			forward_asin_op(p, q, i_var, size_t(arg[0]), J, taylor);
 			break;
 			// -------------------------------------------------
 
@@ -379,7 +379,7 @@ void forward1(
 			case AsinhOp:
 			// sqrt(1 + x * x), asinh(x)
 			CPPAD_ASSERT_UNKNOWN( i_var < numvar  );
-			forward_asinh_op(p, q, i_var, arg[0], J, taylor);
+			forward_asinh_op(p, q, i_var, size_t(arg[0]), J, taylor);
 			break;
 # endif
 			// -------------------------------------------------
@@ -387,7 +387,7 @@ void forward1(
 			case AtanOp:
 			// 1 + x * x, atan(x)
 			CPPAD_ASSERT_UNKNOWN( i_var < numvar  );
-			forward_atan_op(p, q, i_var, arg[0], J, taylor);
+			forward_atan_op(p, q, i_var, size_t(arg[0]), J, taylor);
 			break;
 			// -------------------------------------------------
 
@@ -395,7 +395,7 @@ void forward1(
 			case AtanhOp:
 			// 1 - x * x, atanh(x)
 			CPPAD_ASSERT_UNKNOWN( i_var < numvar  );
-			forward_atanh_op(p, q, i_var, arg[0], J, taylor);
+			forward_atanh_op(p, q, i_var, size_t(arg[0]), J, taylor);
 			break;
 # endif
 			// -------------------------------------------------
@@ -410,14 +410,14 @@ void forward1(
 			case CosOp:
 			// sin(x), cos(x)
 			CPPAD_ASSERT_UNKNOWN( i_var < numvar  );
-			forward_cos_op(p, q, i_var, arg[0], J, taylor);
+			forward_cos_op(p, q, i_var, size_t(arg[0]), J, taylor);
 			break;
 			// ---------------------------------------------------
 
 			case CoshOp:
 			// sinh(x), cosh(x)
 			CPPAD_ASSERT_UNKNOWN( i_var < numvar  );
-			forward_cosh_op(p, q, i_var, arg[0], J, taylor);
+			forward_cosh_op(p, q, i_var, size_t(arg[0]), J, taylor);
 			break;
 			// -------------------------------------------------
 
@@ -510,13 +510,13 @@ void forward1(
 			// -------------------------------------------------
 
 			case ExpOp:
-			forward_exp_op(p, q, i_var, arg[0], J, taylor);
+			forward_exp_op(p, q, i_var, size_t(arg[0]), J, taylor);
 			break;
 			// ---------------------------------------------------
 
 # if CPPAD_USE_CPLUSPLUS_2011
 			case Expm1Op:
-			forward_expm1_op(p, q, i_var, arg[0], J, taylor);
+			forward_expm1_op(p, q, i_var, size_t(arg[0]), J, taylor);
 			break;
 # endif
 			// ---------------------------------------------------
@@ -653,13 +653,13 @@ void forward1(
 			// -------------------------------------------------
 
 			case LogOp:
-			forward_log_op(p, q, i_var, arg[0], J, taylor);
+			forward_log_op(p, q, i_var, size_t(arg[0]), J, taylor);
 			break;
 			// -------------------------------------------------
 
 # if CPPAD_USE_CPLUSPLUS_2011
 			case Log1pOp:
-			forward_log1p_op(p, q, i_var, arg[0], J, taylor);
+			forward_log1p_op(p, q, i_var, size_t(arg[0]), J, taylor);
 			break;
 # endif
 			// -------------------------------------------------
@@ -795,26 +795,26 @@ void forward1(
 			case SignOp:
 			// sign(x)
 			CPPAD_ASSERT_UNKNOWN( i_var < numvar  );
-			forward_sign_op(p, q, i_var, arg[0], J, taylor);
+			forward_sign_op(p, q, i_var, size_t(arg[0]), J, taylor);
 			break;
 			// -------------------------------------------------
 
 			case SinOp:
 			// cos(x), sin(x)
 			CPPAD_ASSERT_UNKNOWN( i_var < numvar  );
-			forward_sin_op(p, q, i_var, arg[0], J, taylor);
+			forward_sin_op(p, q, i_var, size_t(arg[0]), J, taylor);
 			break;
 			// -------------------------------------------------
 
 			case SinhOp:
 			// cosh(x), sinh(x)
 			CPPAD_ASSERT_UNKNOWN( i_var < numvar  );
-			forward_sinh_op(p, q, i_var, arg[0], J, taylor);
+			forward_sinh_op(p, q, i_var, size_t(arg[0]), J, taylor);
 			break;
 			// -------------------------------------------------
 
 			case SqrtOp:
-			forward_sqrt_op(p, q, i_var, arg[0], J, taylor);
+			forward_sqrt_op(p, q, i_var, size_t(arg[0]), J, taylor);
 			break;
 			// -------------------------------------------------
 
@@ -898,14 +898,14 @@ void forward1(
 			case TanOp:
 			// tan(x)^2, tan(x)
 			CPPAD_ASSERT_UNKNOWN( i_var < numvar  );
-			forward_tan_op(p, q, i_var, arg[0], J, taylor);
+			forward_tan_op(p, q, i_var, size_t(arg[0]), J, taylor);
 			break;
 			// -------------------------------------------------
 
 			case TanhOp:
 			// tanh(x)^2, tanh(x)
 			CPPAD_ASSERT_UNKNOWN( i_var < numvar  );
-			forward_tanh_op(p, q, i_var, arg[0], J, taylor);
+			forward_tanh_op(p, q, i_var, size_t(arg[0]), J, taylor);
 			break;
 			// -------------------------------------------------
