+++ cppad/local/sweep/forward2.hpp
@@ -243,7 +243,7 @@ void forward2(
 		switch( op )
 		{
 			case AbsOp:
-			forward_abs_op_dir(q, r, i_var, arg[0], J, taylor);
+			forward_abs_op_dir(q, r, i_var, size_t(arg[0]), J, taylor);
 			break;
 			// -------------------------------------------------
 
@@ -261,7 +261,7 @@ void forward2(
 			case AcosOp:
 			// sqrt(1 - x * x), acos(x)
 			CPPAD_ASSERT_UNKNOWN( i_var < numvar  );
-			forward_acos_op_dir(q, r, i_var, arg[0], J, taylor);
+			forward_acos_op_dir(q, r, i_var, size_t(arg[0]), J, taylor);
 			break;
 			// -------------------------------------------------
 
@@ -269,7 +269,7 @@ void forward2(
 			case AcoshOp:
 			// sqrt(x * x - 1), acosh(x)
 			CPPAD_ASSERT_UNKNOWN( i_var < numvar  );
-			forward_acosh_op_dir(q, r, i_var, arg[0], J, taylor);
+			forward_acosh_op_dir(q, r, i_var, size_t(arg[0]), J, taylor);
 			break;
 # endif
 			// -------------------------------------------------
@@ -277,7 +277,7 @@ void forward2(
 			case AsinOp:
 			// sqrt(1 - x * x), asin(x)
 			CPPAD_ASSERT_UNKNOWN( i_var < numvar  );
-			forward_asin_op_dir(q, r, i_var, arg[0], J, taylor);
+			forward_asin_op_dir(q, r, i_var, size_t(arg[0]), J, taylor);
 			break;
 			// -------------------------------------------------
 
@@ -285,7 +285,7 @@ void forward2(
 			case AsinhOp:
 			// sqrt(1 + x * x), asinh(x)
 			CPPAD_ASSERT_UNKNOWN( i_var < numvar  );
-			forward_asinh_op_dir(q, r, i_var, arg[0], J, taylor);
+			forward_asinh_op_dir(q, r, i_var, size_t(arg[0]), J, taylor);
 			break;
 # endif
 			// -------------------------------------------------
@@ -293,7 +293,7 @@ void forward2(
 			case AtanOp:
 			// 1 + x * x, atan(x)
 			CPPAD_ASSERT_UNKNOWN( i_var < numvar  );
-			forward_atan_op_dir(q, r, i_var, arg[0], J, taylor);
+			forward_atan_op_dir(q, r, i_var, size_t(arg[0]), J, taylor);
 			break;
 			// -------------------------------------------------
 
@@ -301,7 +301,7 @@ void forward2(
 			case AtanhOp:
 			// 1 - x * x, atanh(x)
 			CPPAD_ASSERT_UNKNOWN( i_var < numvar  );
-			forward_atanh_op_dir(q, r, i_var, arg[0], J, taylor);
+			forward_atanh_op_dir(q, r, i_var, size_t(arg[0]), J, taylor);
 			break;
 # endif
 			// -------------------------------------------------
@@ -316,14 +316,14 @@ void forward2(
 			case CosOp:
 			// sin(x), cos(x)
 			CPPAD_ASSERT_UNKNOWN( i_var < numvar  );
-			forward_cos_op_dir(q, r, i_var, arg[0], J, taylor);
+			forward_cos_op_dir(q, r, i_var, size_t(arg[0]), J, taylor);
 			break;
 			// ---------------------------------------------------
 
 			case CoshOp:
 			// sinh(x), cosh(x)
 			CPPAD_ASSERT_UNKNOWN( i_var < numvar  );
-			forward_cosh_op_dir(q, r, i_var, arg[0], J, taylor);
+			forward_cosh_op_dir(q, r, i_var, size_t(arg[0]), J, taylor);
 			break;
 			// -------------------------------------------------
 
@@ -379,13 +379,13 @@ void forward2(
 # endif
 
 			case ExpOp:
-			forward_exp_op_dir(q, r, i_var, arg[0], J, taylor);
+			forward_exp_op_dir(q, r, i_var, size_t(arg[0]), J, taylor);
 			break;
 			// -------------------------------------------------
 
 # if CPPAD_USE_CPLUSPLUS_2011
 			case Expm1Op:
-			forward_expm1_op_dir(q, r, i_var, arg[0], J, taylor);
+			forward_expm1_op_dir(q, r, i_var, size_t(arg[0]), J, taylor);
 			break;
 # endif
 			// -------------------------------------------------
@@ -431,13 +431,13 @@ void forward2(
 			// -------------------------------------------------
 
 			case LogOp:
-			forward_log_op_dir(q, r, i_var, arg[0], J, taylor);
+			forward_log_op_dir(q, r, i_var, size_t(arg[0]), J, taylor);
 			break;
 			// ---------------------------------------------------
 
 # if CPPAD_USE_CPLUSPLUS_2011
 			case Log1pOp:
-			forward_log1p_op_dir(q, r, i_var, arg[0], J, taylor);
+			forward_log1p_op_dir(q, r, i_var, size_t(arg[0]), J, taylor);
 			break;
 # endif
 			// ---------------------------------------------------
@@ -485,26 +485,26 @@ void forward2(
 			case SignOp:
 			// sign(x)
 			CPPAD_ASSERT_UNKNOWN( i_var < numvar  );
-			forward_sign_op_dir(q, r, i_var, arg[0], J, taylor);
+			forward_sign_op_dir(q, r, i_var, size_t(arg[0]), J, taylor);
 			break;
 			// -------------------------------------------------
 
 			case SinOp:
 			// cos(x), sin(x)
 			CPPAD_ASSERT_UNKNOWN( i_var < numvar  );
-			forward_sin_op_dir(q, r, i_var, arg[0], J, taylor);
+			forward_sin_op_dir(q, r, i_var, size_t(arg[0]), J, taylor);
 			break;
 			// -------------------------------------------------
 
 			case SinhOp:
 			// cosh(x), sinh(x)
 			CPPAD_ASSERT_UNKNOWN( i_var < numvar  );
-			forward_sinh_op_dir(q, r, i_var, arg[0], J, taylor);
+			forward_sinh_op_dir(q, r, i_var, size_t(arg[0]), J, taylor);
 			break;
 			// -------------------------------------------------
 
 			case SqrtOp:
-			forward_sqrt_op_dir(q, r, i_var, arg[0], J, taylor);
+			forward_sqrt_op_dir(q, r, i_var, size_t(arg[0]), J, taylor);
 			break;
 			// -------------------------------------------------
 
@@ -536,14 +536,14 @@ void forward2(
 			case TanOp:
 			// tan(x)^2, tan(x)
 			CPPAD_ASSERT_UNKNOWN( i_var < numvar  );
-			forward_tan_op_dir(q, r, i_var, arg[0], J, taylor);
+			forward_tan_op_dir(q, r, i_var, size_t(arg[0]), J, taylor);
 			break;
 			// -------------------------------------------------
 
 			case TanhOp:
 			// tanh(x)^2, tanh(x)
 			CPPAD_ASSERT_UNKNOWN( i_var < numvar  );
-			forward_tanh_op_dir(q, r, i_var, arg[0], J, taylor);
+			forward_tanh_op_dir(q, r, i_var, size_t(arg[0]), J, taylor);
 			break;
 			// -------------------------------------------------
